+++ benchmarks/lfstackbench.c
@@ -29,7 +29,7 @@ inline static void *Execute(void *Arg) {
         d1 = getTimeMillis();
 
     for (i = 0; i < bench_args.runs; i++) {
-        LFStackPush(&stack, th_state, id + 1);
+        LFStackPush(&stack, th_state, id);
         rnum = fastRandomRange(1, bench_args.max_work);
         for (j = 0; j < rnum; j++)
             ;
@@ -56,14 +56,14 @@ int main(int argc, char *argv[]) {
     printStats(bench_args.nthreads, bench_args.total_runs);
 
 #ifdef DEBUG
-    int counter = 0;
+    long counter = 0;
 
     while (stack.top != null) {
         counter++;
         stack.top = stack.top->next;
     }
 
-    fprintf(stderr, "DEBUG: %d nodes were left in the stack!\n", counter);
+    fprintf(stderr, "DEBUG: %ld nodes were left in the stack\n", counter);
 #endif
 
     return 0;