+++ libconcurrent/concurrent/dsmqueue.c
@@ -3,13 +3,12 @@
 inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid);
 inline static RetVal serialDequeue(void *state, ArgVal arg, int pid);
 
-static const int GUARD = INT_MIN;
 static __thread PoolStruct pool_node CACHE_ALIGN;
 
 void DSMQueueStructInit(DSMQueueStruct *queue_object_struct, uint32_t nthreads) {
     DSMSynchStructInit(&queue_object_struct->enqueue_struct, nthreads);
     DSMSynchStructInit(&queue_object_struct->dequeue_struct, nthreads);
-    queue_object_struct->guard.val = GUARD;
+    queue_object_struct->guard.val = GUARD_VALUE;
     queue_object_struct->guard.next = null;
     queue_object_struct->first = &queue_object_struct->guard;
     queue_object_struct->last = &queue_object_struct->guard;
@@ -35,13 +34,16 @@ inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid) {
 
 inline static RetVal serialDequeue(void *state, ArgVal arg, int pid) {
     DSMQueueStruct *st = (DSMQueueStruct *)state;
-    Node *node = (Node *)st->first;
+    volatile Node *node, *prev;
 
-    if (st->first->next != null) {
-        RetVal ret = node->val;
+    if (st->first->next != null){
+        prev = st->first;
         st->first = st->first->next;
-        recycle_obj(&pool_node, (void *)node);
-        return ret;
+        node = st->first;
+        if (node->val == GUARD_VALUE)
+            return serialDequeue(state, arg, pid);
+        recycle_obj(&pool_node, (Node *)prev);
+        return node->val;
     } else {
         return -1;
     }