+++ libconcurrent/concurrent/dsmstack.c
@@ -7,7 +7,7 @@ static __thread PoolStruct pool_node CACHE_ALIGN;
 
 void DSMSStackInit(DSMStackStruct *stack_object_struct, uint32_t nthreads) {
     DSMSynchStructInit(&stack_object_struct->object_struct, nthreads);
-    stack_object_struct->head = null;
+    stack_object_struct->top = null;
     StoreFence();
 }
 
@@ -19,11 +19,11 @@ void DSMStackThreadStateInit(DSMStackStruct *object_struct, DSMStackThreadState 
 inline static RetVal serialPushPop(void *state, ArgVal arg, int pid) {
     if (arg == POP_OP) {
         volatile DSMStackStruct *st = (DSMStackStruct *)state;
-        volatile Node *node = st->head;
+        volatile Node *node = st->top;
 
-        if (st->head != null) {
+        if (st->top != null) {
             RetVal ret = node->val;
-            st->head = st->head->next;
+            st->top = st->top->next;
             recycle_obj(&pool_node, (void *)node);
             return ret;
         } else
@@ -33,9 +33,9 @@ inline static RetVal serialPushPop(void *state, ArgVal arg, int pid) {
         Node *node;
 
         node = alloc_obj(&pool_node);
-        node->next = st->head;
+        node->next = st->top;
         node->val = arg;
-        st->head = node;
+        st->top = node;
 
         return 0;
     }
@@ -45,6 +45,6 @@ void DSMStackPush(DSMStackStruct *object_struct, DSMStackThreadState *lobject_st
     DSMSynchApplyOp(&object_struct->object_struct, &lobject_struct->th_state, serialPushPop, object_struct, (ArgVal)arg, pid);
 }
 
-void DSMStackPop(DSMStackStruct *object_struct, DSMStackThreadState *lobject_struct, int pid) {
-    DSMSynchApplyOp(&object_struct->object_struct, &lobject_struct->th_state, serialPushPop, object_struct, (ArgVal)POP_OP, pid);
+RetVal DSMStackPop(DSMStackStruct *object_struct, DSMStackThreadState *lobject_struct, int pid) {
+    return DSMSynchApplyOp(&object_struct->object_struct, &lobject_struct->th_state, serialPushPop, object_struct, (ArgVal)POP_OP, pid);
 }