+++ libconcurrent/concurrent/hstack.c
@@ -7,7 +7,7 @@ static __thread PoolStruct pool_node CACHE_ALIGN;
 
 void HStackInit(HStackStruct *stack_object_struct, uint32_t nthreads, uint32_t numa_nodes) {
     HSynchStructInit(&stack_object_struct->object_struct, nthreads, numa_nodes);
-    stack_object_struct->head = null;
+    stack_object_struct->top = null;
 }
 
 void HStackThreadStateInit(HStackStruct *object_struct, HStackThreadState *lobject_struct, int pid) {
@@ -18,14 +18,14 @@ void HStackThreadStateInit(HStackStruct *object_struct, HStackThreadState *lobje
 inline RetVal serialPushPop(void *state, ArgVal arg, int pid) {
     if (arg == POP_OP) {
         volatile HStackStruct *st = (HStackStruct *)state;
-        volatile Node *node = st->head;
+        volatile Node *node = st->top;
 
-        if (st->head == null) {
+        if (st->top == null) {
             return -1;
         } else {
             RetVal ret = node->val;
 
-            st->head = st->head->next;
+            st->top = st->top->next;
             recycle_obj(&pool_node, (void *)node);
 
             return ret;
@@ -35,9 +35,9 @@ inline RetVal serialPushPop(void *state, ArgVal arg, int pid) {
         Node *node;
 
         node = alloc_obj(&pool_node);
-        node->next = st->head;
+        node->next = st->top;
         node->val = arg;
-        st->head = node;
+        st->top = node;
 
         return 0;
     }