+++ libconcurrent/concurrent/osciqueue.c
@@ -1,15 +1,15 @@
 #include <osciqueue.h>
+#include <queue-stack.h>
 
 inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid);
 inline static RetVal serialDequeue(void *state, ArgVal arg, int pid);
 
-static const int GUARD = INT_MIN;
 
 void OsciQueueInit(OsciQueueStruct *queue_object_struct, uint32_t nthreads, uint32_t fibers_per_thread) {
     OsciInit(&(queue_object_struct->enqueue_struct), nthreads, fibers_per_thread);
     OsciInit(&queue_object_struct->dequeue_struct, nthreads, fibers_per_thread);
     queue_object_struct->pool_node = getAlignedMemory(CACHE_LINE_SIZE, queue_object_struct->enqueue_struct.groups_of_fibers * sizeof(PoolStruct));
-    queue_object_struct->guard.val = GUARD;
+    queue_object_struct->guard.val = GUARD_VALUE;
     queue_object_struct->guard.next = null;
     queue_object_struct->first = &queue_object_struct->guard;
     queue_object_struct->last = &queue_object_struct->guard;
@@ -35,13 +35,16 @@ inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid) {
 
 inline static RetVal serialDequeue(void *state, ArgVal arg, int pid) {
     OsciQueueStruct *st = (OsciQueueStruct *)state;
-    Node *node = (Node *)st->first;
+    volatile Node *node, *prev;
 
-    if (st->first->next != null) {
-        RetVal ret = node->val;
+    if (st->first->next != null){
+        prev = st->first;
         st->first = st->first->next;
-        recycle_obj(&(st->pool_node[getThreadId()]), (void *)node);
-        return ret;
+        node = st->first;
+        if (node->val == GUARD_VALUE)
+            return serialDequeue(state, arg, pid);
+        recycle_obj(&(st->pool_node[getThreadId()]), (Node *)prev);
+        return node->val;
     } else {
         return -1;
     }