+++ libconcurrent/concurrent/oscistack.c
@@ -7,7 +7,7 @@ static const int POP_OP = INT_MIN;
 void OsciStackInit(OsciStackStruct *stack_object_struct, uint32_t nthreads, uint32_t fibers_per_thread) {
     OsciInit(&(stack_object_struct->object_struct), nthreads, fibers_per_thread);
     stack_object_struct->pool_node = getAlignedMemory(CACHE_LINE_SIZE, stack_object_struct->object_struct.groups_of_fibers * sizeof(PoolStruct));
-    stack_object_struct->head = null;
+    stack_object_struct->top = null;
 }
 
 void OsciStackThreadStateInit(OsciStackStruct *object_struct, OsciStackThreadState *lobject_struct, int pid) {
@@ -18,11 +18,11 @@ void OsciStackThreadStateInit(OsciStackStruct *object_struct, OsciStackThreadSta
 inline static RetVal serialPushPop(void *state, ArgVal arg, int pid) {
     if (arg == POP_OP) {
         volatile OsciStackStruct *st = (OsciStackStruct *)state;
-        volatile Node *node = st->head;
+        volatile Node *node = st->top;
 
-        if (st->head != null) {
+        if (st->top != null) {
             RetVal ret = node->val;
-            st->head = st->head->next;
+            st->top = st->top->next;
             recycle_obj(&(st->pool_node[getThreadId()]), (void *)node);
             return ret;
         }
@@ -33,9 +33,9 @@ inline static RetVal serialPushPop(void *state, ArgVal arg, int pid) {
         Node *node;
 
         node = alloc_obj(&(st->pool_node[getThreadId()]));
-        node->next = st->head;
+        node->next = st->top;
         node->val = arg;
-        st->head = node;
+        st->top = node;
 
         return 0;
     }
@@ -45,6 +45,6 @@ void OsciStackApplyPush(OsciStackStruct *object_struct, OsciStackThreadState *lo
     OsciApplyOp(&(object_struct->object_struct), &(lobject_struct->th_state), serialPushPop, object_struct, (ArgVal)arg, pid);
 }
 
-void OsciStackApplyPop(OsciStackStruct *object_struct, OsciStackThreadState *lobject_struct, int pid) {
-    OsciApplyOp(&(object_struct->object_struct), &(lobject_struct->th_state), serialPushPop, object_struct, POP_OP, pid);
+RetVal OsciStackApplyPop(OsciStackStruct *object_struct, OsciStackThreadState *lobject_struct, int pid) {
+    return OsciApplyOp(&(object_struct->object_struct), &(lobject_struct->th_state), serialPushPop, object_struct, POP_OP, pid);
 }